[cleaner] Completed properties 1,6
Spectra - property 1:
- Added sys variable robot_rotation to describe robot's appearance.
- Added guaranty to make sure the robot starts its rotation from 0
degrees when reaches the orange zone.
- Added guaranty to make sure the robot repeatedly switches its usual
appearance (rotates 90 degrees to the right at each step) when it
reaches the orange zone until it leaves it.
- Added predicate to describe when the robot is at orange zone.

Spectra - property 6:
- Added assumption that greenLight always eventually turns on to make
sure guaranty in line 26 is not violated.
- Added guaranty to make sure robot stays at origin in the first 8
states of its execution and then starts moving only when there is green
light.