[cleaner] Added property 5 spec and updated Java code
Spectra:
- Added sys var 'no_target_counter' to count the consecutive steps
without a cleaning request.
- Added sys var 'sleep_mode' to indicate if the robot is sleeping or
not.
- Added YELLOW and GRAY colors and updated color transitions according
to the requirements.
- Added guarantees for 'no_target_counter' transitions using predicates
for inc, reset, enter sleep, wake up (see code for detailed
documentation).
- Added gar 'alwEv cleaning_target' instead of the trig which made the
robot reach a target after a cleaning request and the pattern which
caused the robot visit origin after a tank is full (due to possible bug
in the pattern implementation).

Java:
- Updated java code for handling the env variables and the new colors.

Simulation:
- Run accordingly to the requirements.
- Although GUI simulation will probably never reach a state where the
robot is in sleep mode and GRAY, the walker does show it if the env
defines 8 consecutive states without a cleaning request.