[junction] Updated basic gars and asms and added emergency priorities
- Removed triggers and used simple implies clauses.
- Added assumption that alwEv there will be traffic in any lane or side
walk and also that alwEv there won't be a traffic. This enabled to
replace the triggers that caused a green light to turn on after a
pedestrian/vehicle arrive with simple alwEv clause for
pedestrian/vehicle and green light. In addition to simplifying the code,
it made the realizability check faster.
- Added emergency priorities - documentation inside the code.
- Moved unrealizable code to deprecated section (should be removed
later).