[junction] Added 'mode' sys variable to prevent traffic starvation
- Removed assumptions that caused the specification to be non well
separated - i.e the asm for alwEv traffic and alwEv no traffic.
- Removed guarantees for alwEv green light (after removing the
assumptions above it was necessary due to unrealizability). After
removing this there is a possibily for a traffic starvation.
- Added new sys var 'mode' to allow prevent the traffic starvation - if
mode is 'AUTO' then the sys behaves as normal, if the mode is
'PEDESTRIANS' then pedestrians have a priority and same for north
vehicles and south vehicles.
- Added a counter and guarantees to switch between the modes every
several states and thus prevent traffic starvation (there still can be
starvation if there is fog or freeze mode but it reduces the probability
for starvation considerably).