[junction][extended] Added manual mode
Spectra:
- Added env variables 'manualTrafficLight' and 'trafficLightColor' to
indicate the environment manual modes' choice.

- Added guarantee to ensure that if manual mode is on then the specific
traffic light it controls must turn to the color that the environment
chooses.

- Added assumptions that manual mode cannot occur at the same as fog or
freeze mode (because it can violate the requirements of the task).

- Added assumption that when there are road constructions, manual mode
cannot control the south straight lane vehicle's traffic light.

- All basic guarantees and logic were updated to support manual mode
logic - the guarantees remain valid as long as manual mode is off or the
traffic light that manual mode controls is not the traffic light that
the guarantees handle.

- The cyclic logic for starvation prevention is not applied when manual
mode is on (i.e only 'AUTO' mode is allowed).

- Updated emergency vehicles priority logic so that when manual mode is
on, all emergency priorities must comply with the traffic light that is
controlled by the manual mode. For example, an emergency vehicle might
not get a priority when manual mode is on and forces one of the traffic
light that interfere with it's path to become GREEN. It also will not
get a priority if manual mode forces it to become red. When manual mode
is off, all regular priority rules apply.

Java:
- Added manual mode option in the GUI.
- Added image for manual mode.

Simulation:
- All requirements have been tested in the simulation and it works as
expected.